requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
cloudinary==1.29.0
cryptography==37.0.2
Django==4.0.4
django-phonenumber-field==6.3.0
djangorestframework==3.13.1
jwt==1.3.1
phonenumbers==8.12.51
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
